Refactored `AttestationIssuer` to separate concerns into dedicated files and updated references accordingly.
