+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/adapter/input/web/security/DPoPNonce.kt
@@ -17,8 +17,8 @@ package eu.europa.ec.eudi.pidissuer.adapter.input.web.security
 
 import arrow.core.NonEmptySet
 import com.nimbusds.jose.JWSAlgorithm
-import eu.europa.ec.eudi.pidissuer.port.out.credential.GenerateNonce
-import eu.europa.ec.eudi.pidissuer.port.out.credential.VerifyNonce
+import eu.europa.ec.eudi.pidissuer.port.out.attestation.GenerateNonce
+import eu.europa.ec.eudi.pidissuer.port.out.attestation.VerifyNonce
 import kotlin.time.Duration
 import kotlin.time.Instant
