+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/domain/CredentialIssuerMetaData.kt
@@ -25,7 +25,7 @@ import com.nimbusds.jose.jwk.ECKey
 import com.nimbusds.jose.jwk.JWKSet
 import com.nimbusds.jose.jwk.RSAKey
 import eu.europa.ec.eudi.pidissuer.domain.OpenId4VciSpec.ZIP_ALGORITHMS
-import eu.europa.ec.eudi.pidissuer.port.out.AttestationIssuer
+import eu.europa.ec.eudi.pidissuer.port.out.attestation.AttestationIssuer
 import java.util.*
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.days
@@ -250,7 +250,7 @@ data class CredentialIssuerMetaData(
             "credential configuration ids must be unique"
         }
 
-        val maxValidityOfIssuedCredentials = attestationIssuers.maxOf { it.validity }
+        val maxValidityOfIssuedCredentials = attestationIssuers.maxOf { it.configuration.validity }
 
         require(preferredClientStatusPeriod.value >= maxValidityOfIssuedCredentials) {
             "preferredClientStatusPeriod must be greater than or equal to the max validity of issued credentials configured"
