+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/domain/SdJwtVcProfile.kt
@@ -20,6 +20,7 @@ import arrow.core.nonEmptySetOf
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.jwk.JWK
 import eu.europa.ec.eudi.sdjwt.SdJwtVcSpec
+import kotlin.time.Duration
 
 const val SD_JWT_VC_FORMAT_VALUE = SdJwtVcSpec.MEDIA_SUBTYPE_DC_SD_JWT
 val SD_JWT_VC_FORMAT = Format(SD_JWT_VC_FORMAT_VALUE)
@@ -34,16 +35,21 @@ value class SdJwtVcType(
  */
 data class SdJwtVcCredentialConfiguration(
     override val id: CredentialConfigurationId,
-    val type: SdJwtVcType,
     override val scope: Scope,
-    val credentialSigningAlgorithmsSupported: NonEmptySet<JWSAlgorithm>?,
-    val publicKey: JWK,
     override val display: List<CredentialDisplay>,
-    val claims: List<ClaimDefinition>,
     override val deviceBinding: DeviceBinding,
     override val attestationCategory: AttestationCategory,
     override val credentialReusePolicy: CredentialReusePolicy = CredentialReusePolicy.None,
+    override val validity: Duration,
+    val type: SdJwtVcType,
+    val credentialSigningAlgorithmsSupported: NonEmptySet<JWSAlgorithm>?,
+    val publicKey: JWK,
+    val claims: List<ClaimDefinition>,
 ) : CredentialConfiguration {
+    init {
+        require(validity.isPositive()) { "'validity' must be a positive duration" }
+    }
+
     override val cryptographicBindingMethodsSupported: NonEmptySet<CryptographicBindingMethod>?
         get() =
             when (deviceBinding) {
